+++ src/elements/GeoBorderedBox/GeoBorderedBoxHeaderSearchForm.vue
@@ -4,14 +4,33 @@
     @submit.prevent
   >
     <geo-input
-      :leading-accessory-icon="searchIcon"
       :value="value"
       :placeholder="placeholder"
       v-bind="$attrs"
       type="text"
       @input="searchPattern($event)"
       @delete-value="deleteValue($event)"
-    />
+    >
+      <!-- @slot Use this slot to customize content displayed before the icon -->
+      <slot
+        slot="leadingAccessoryItem"
+        name="beforeSearchIconItem"
+      />
+      <font-awesome-icon
+        slot="leadingAccessoryItem"
+        :icon="searchIcon"
+      />
+      <!-- @slot Use this slot to customize content displayed after the icon, on the left of the input -->
+      <slot
+        slot="leadingAccessoryItem"
+        name="afterSearchIconItem"
+      />
+      <!-- @slot Use this slot to customize content displayed on the right of the input -->
+      <slot
+        slot="trailingAccessoryItem"
+        name="trailingAccessoryItem"
+      />
+    </geo-input>
   </form>
 </template>