+++ src/elements/GeoInput/GeoInputPrefix.vue
@@ -1,0 +1,17 @@
+<template>
+  <div :class="`geo-input-prefix${cssSuffix}`">
+    <!-- @slot Use this slot to customize what's displayed as a prefix -->
+    <slot />
+  </div>
+</template>
+
+<script>
+import cssSuffix from '../../mixins/cssModifierMixin'
+
+export default {
+  name: 'GeoInputPrefix',
+  status: 'ready',
+  release: '27.0.0',
+  mixins: [cssSuffix]
+}
+</script>