+++ src/elements/GeoSelect/GeoSelectToggleButton.vue
@@ -3,7 +3,8 @@
     :class="{
       [`geo-select-toggle-button${cssSuffix}`]: true,
       [`geo-select-toggle-button--empty${cssSuffix}`]: isEmpty,
-      [`geo-select-toggle-button--disabled${cssSuffix}`]: disabled
+      [`geo-select-toggle-button--disabled${cssSuffix}`]: disabled,
+      [`geo-select-toggle-button--${variant}${cssSuffix}`]: true
     }"
     @click="handleClick($event)"
   >
@@ -32,10 +33,18 @@
 <script>
 import cssSuffix from '../../mixins/cssModifierMixin'
 
+export const VARIANTS = {
+  inputAccessorySuffix: 'inputAccessorySuffix',
+  inputAccessoryPrefix: 'inputAccessoryPrefix'
+}
+
 export default {
   name: 'GeoSelectToggleButton',
   status: 'ready',
   release: '4.1.0',
+  constants: {
+    VARIANTS
+  },
   mixins: [cssSuffix],
   props: {
     /**
@@ -108,6 +117,30 @@ export default {
     disabled: {
       type: Boolean,
       default: false
+    },
+
+    /**
+     * Predefined style scheme of the toggle button, allowing the possibility to use
+     * it as regular or as a suffix of an input.
+     *
+     * Supported `variant` values are exported under `VARIANTS` named export.
+     * See [Component Constants](/docs/components-constants.html) for more info on how
+     * to use those constants in your code.
+     *
+     * > **Note:** You can always override the color scheme of any
+     * > `GeoSelectToggleButton` using `cssModifier` prop.
+     */
+    variant: {
+      type: String,
+      required: false,
+      validator (value) {
+        if (!value) return true
+        if (value in VARIANTS) return true
+
+        const supportedValues = Object.values(VARIANTS).map(i => `«${i}»`).join(', ')
+        console.warn(`GeoSelectToggleButton [component] :: Unsupported value («${value}») for «variant» property. Use one of ${supportedValues}`)
+        return false
+      }
     }
   },
   computed: {