+++ CHANGELOG.md
@@ -14,6 +14,18 @@ Fixed:
 
 - Import `date-fns` functions separately
 
+## 27.0.0
+
+Breaking:
+
+- `GeoInput` renamed `suffix` and `prefix` slots to `leadingAccessoryItem` and `trailingAccessoryItem`
+
+New:
+
+- `GeoInputSuffix` component
+- `GeoInputPrefix` component
+- `GeoSelectToggleButton` can have variant `inputAccessorySuffix` and `inputAccessoryPrefix`
+
 ## 26.2.0
 
 New: